Add files via upload
🎉 Just completed my 4th task during my internship at CodSoft ! 🚀 Created a Rock-Paper-Scissors game ✊🖐✌️ that’s both fun and challenging. Excited to continue learning and growing through these amazing experiences! 💪 

° Rock-Paper-Scissors Game -

This Python code implements a Rock-Paper-Scissors game where the user plays against the computer. It prompts the user to choose "rock," "paper," or "scissors," determines the winner, and keeps track of the scores. The game continues in a loop until the user decides to exit by typing "exit" or chooses not to play another round.